externalize debounce func
