+++ app/javascript/lib/shared/modules/debounce.js
@@ -1,0 +1,10 @@
+export const debounce = (func, timeout) => {
+  let timer = undefined;
+  return (...args) => {
+    const next = () => func(...args);
+    if (timer) {
+      clearTimeout(timer);
+    }
+    timer = setTimeout(next, timeout > 0 ? timeout : 300);
+  };
+}
